feat : useWeeklyStatusModel ->useFormatModel 변경 및 type 수정
