+++ src/components/TotalAdStatus.tsx
@@ -4,16 +4,23 @@ import { Item } from "../styles/Item";
 import { useTotalAdStatusModel } from "../models/useTotalAdStatusModel";
 import { useRecoilState } from "recoil";
 import { totalAdStatusState } from "../store/atom";
-import { TotalAdStatusType } from "../types/totalAdStatusType";
-import { LineChart, Line, CartesianGrid, XAxis, YAxis, Legend } from "recharts";
-import { calcWeeklySum, useWeeklyStatusCallback } from "../models/useWeeklyStatusModel";
+import { TotalAdStatusType } from "../models/types/index";
+import {
+  LineChart,
+  Line,
+  CartesianGrid,
+  XAxis,
+  YAxis,
+  ResponsiveContainer,
+} from "recharts";
+import { calculateSum, calculateSumCallback } from "../models/useFormatModel";
+import Brightness1Icon from "@mui/icons-material/Brightness1";
 
-//any, dependency 처리
-//차트 브라우저 width에 따라 변하게 해야함
+//TODO dependency 처리, map key 처리
 const TotalAdStatus = () => {
   const [totalAdStatus, setTotalAdStatus] =
     useRecoilState<TotalAdStatusType[]>(totalAdStatusState);
-    const [weeklyData, setWeeklyData] = React.useState<string[]>([]);
+  const [weeklyData, setWeeklyData] = React.useState<string[]>([]);
 
   //일주일치 데이터 하드 코딩 2022-02-01 ~ 2022-02-07
   // "?date_gte=2022-02-01&date_lte=2022-02-07"
@@ -26,9 +33,10 @@ const TotalAdStatus = () => {
       .catch((error) => console.log("data fetching error! detail:", error));
   }, []);
 
-  React.useEffect(()=>{
-    setWeeklyData(calcWeeklySum(totalAdStatus, useWeeklyStatusCallback));
-  },[totalAdStatus]);
+  React.useEffect(
+    () => setWeeklyData(calculateSum(totalAdStatus, calculateSumCallback)),
+    [totalAdStatus]
+  );
 
   return (
     <Box sx={{ p: 3 }}>
@@ -50,27 +58,45 @@ const TotalAdStatus = () => {
               justifyContent: "center",
             }}
           >
-            {weeklyData?.map((data : string)=>{
-              return(
-              <Grid item xs={12} md={4}>
-                <Item>{data}</Item>
-              </Grid>);
-              }
-            )}
+            {weeklyData?.map((data: string, index: number) => {
+              return (
+                <Grid key={index} item xs={12} md={4}>
+                  <Item>{data}</Item>
+                </Grid>
+              );
+            })}
           </Grid>
         </Container>
 
-        <Container sx={{ display: "flex", justifyContent: "center", p: 2 }}>
-          <LineChart
-            width={900}
-            height={400}
-            data={totalAdStatus}
-            margin={{ top: 20, right: 20, bottom: 40, left: 10 }}
+        <Container
+          sx={{ display: "flex", alignItems: "center", height: "4rem" }}
+        >
+          <Item sx={{ width: "8rem", display: "flex", alignItems: "center" }}>
+            <Brightness1Icon sx={{ color: "blue", mr: "2rem", fontSize: 10 }} />
+            roas
+          </Item>
+          <Item
+            sx={{
+              width: "8rem",
+              ml: "1rem",
+              display: "flex",
+              alignItems: "center",
+            }}
           >
-            <Legend layout="vertical" align="left" verticalAlign="top" />
+            <Brightness1Icon
+              sx={{ color: "green", mr: "2rem", fontSize: 10 }}
+            />
+            click
+          </Item>
+        </Container>
+
+        <ResponsiveContainer width="100%" aspect={2}>
+          <LineChart
+            data={totalAdStatus}
+            margin={{ top: 40, right: 20, bottom: 40, left: 10 }}
+          >
             <CartesianGrid stroke="#ccc" horizontal={true} vertical={false} />
             <XAxis
-              // dataKey={"name"}
               dataKey={"date"}
               axisLine={false}
               tickLine={false}
@@ -81,7 +107,7 @@ const TotalAdStatus = () => {
             <Line type="monotone" dataKey="roas" stroke="blue" />
             <Line type="monotone" dataKey="click" stroke="green" />
           </LineChart>
-        </Container>
+        </ResponsiveContainer>
       </Container>
     </Box>
   );
