--- src/types/totalAdStatusType.ts
@@ -1,14 +1,0 @@
-export type TotalAdStatusType = {
-  imp: number;
-  click: number;
-  cost: number;
-  conv: number;
-  convValue: number;
-  ctr: number;
-  cvr: number;
-  cpc: number;
-  cpa: number;
-  roas: number;
-  date: string;
-  name?: string;
-};
